The Character Class - Step 3
Created a Character Class (EchoCharacter) to use as a base for the main character in the game. Coded basic movement for the Character including running, jumping and looking around using the new Enhanced Input system for Unreal Engine 5. We don't have actual animation yet, but the Character performs the directional movement as expected. Will start on amination in the next commit.