The Animation Blueprint - Step 4
Created an Anim Instance Class (EchoAnimInstance) to use as a parent class for the animation blueprint (ABP_Echo). So far we just have basic movement including idle, running and jumping. Also built a Control Rig so that the character's feet look natural when standing on inclines. The the knees bend so the feet can adjust to different angles of the surface that the character is standing on. Not a lot of C++ code for this commit but Control Rigs do take a LOT of node sin the blueprint!