The Weapon Class - Step 6
Created the Weapon class as a child of the Item class and added a socket to the right hand of the character's skeleton called RightHandSocket. Created an IK Rig to adjust the character's foot placement when standing on uneven ground. Created an enum class for the character's state to show correct animations depending on whether the character is holding a weapon (equipped) or not (unequipped). Created two additional animation blueprints to control locomotion and IK Rig functions, respectively and linked them to the main animation blueprint. As long as Weapon is being overlapped, the Weapon will attach to the character's hand when EKeyPressed function is called.