Weapon Mechanics - Step 8
Created an Enemy class for the Character to attack in the game. Added a collision box to the Weapon class for tracing. Used the Attack montage to turn the collision box on and off during attacking. Also created an Interface for the Enemy to capture hits by the Character during attacking. Created a Hit React montage for four different animations that the Enemy uses depending on what angle it gets attacked by the Character. Finally added finishing touches for attacks such as hit particles, hit sounds and weapon trails that all happen when the Character attacks the Enemy.