Breakable Actors - Step 9
Created a new Breakable Actor class and coded in a Field System to make it breakable into even pieces on the screen. Changed GetHit function on the interface to a Blueprint Native Event so it can be used both in C++ and/or in blueprints to any Actor that inherits the IHitInterface. Added a Field System to the Weapon, then used the GetHit function in blueprints on the Breakable Actor to generate break sounds when the function is called (i.e. when it is struck by the Weapon). We might be putting some treasure inside one of the actors in the next commit.