Treasure - Step 10
Updated the BreakableActor Class that is now the parent of objects such as pots and vases that the Character can smash into many pieces when hit with the Character's Weapon. Created a Treasure Class is used to parent spawned Items such as gems and gold when the BreakableActor Items are destroyed. I also added a Niagara Component to the Item class. This Niagara effect is displayed over newly spawned Treasure and Weapons. The Niagara effect is deactivated once an item is picked up by the Character.