Combat - Step 11
Created an Actor Component (AttributeComponent) to store stats on a Enemy such as health. Also created a Widget Component (HealthBar) to display the current health percentage of the Enemy during gameplay. Used the built-in Damage/TakeDamage functions for the Character class to apply damage to the Enemy when struck by the Character's Weapon. Customized the HealthBar widget for a more pleasing look. Created a Death Montage with 5 different Death Animations that is picked at random when the Enemy dies. Finally added some finishing touches to the Enemy Death logic so it looks more real when the Enemy dies. Hopefully will add a death sound to the Enemy in a future commit.