Enemy Behavior - Step 12
Created a blendspace that allows the correct animation to play when the Enemy walks, runs or is simply standing idle. Added patrol target actors to the training map that allows Enemies to walk from one point of the map to another. Added a Pawn Sensing Component to the Enemy that allows them to see the Character within a certain sight radius and then chases and "attacks" (attack behavior not yet coded) when in proper range. If the Character gets far enough, the Enemy will lose interest and return to patrolling which is the default behavior. Finally, added some code that will agro the Enemy when attacked from behind. Will focus on Enemy attack behavior in the next commit!