Enemy Attacks - Step 12
Created an Actor (BaseCharacter) class to serve as a base foundation for all Characters in the game including the Player and Enemies. All variables and functions that are commonly used between both Characters, such as hit sounds, react montages, spawned particles, etc., are now derived from the base class. Created an animation montage for the Enemy to attack when in the "Attacking" and "Engaged" states. Updated the Weapon class so that the owner of the Weapon doesn't damage other characters of the same class.

Did a LOT of refactoring in this commit to easily show which functions are derived from parent classes. I also put all of the includes statements in alphabetical order for easier reading.