Smarter Enemies - Step 14
Created an IK Rig for the Paladin and retargeted the four react animations from the Paladin to the Echo character. Echo now reacts to getting hit by Weapons in the same manner as her Enemies. Added motion warping to the Enemy's attack montage to make it slightly more difficult for Echo to dodge or run away during an Enemy attack. The next commit will focus on getting Echo's attributes up and running.