Echo's Attributes - Step 15
Created a default HUD to use on the GameMode that show's the health and stamina bars for Echo. Retargeted and duplicated the death animations to be used on Echo, but probably will be using different ones from Mixamo in the future. Added code to stop the Enemy from attacking Echo after she's dead. Next commit will focus on Souls and Stamina.