Souls & Stamina - Step 16
Created a BP_Soul particle system blueprint from the Niagara System as a child of the newly created Soul class. This particle effect spawns from Enemies at the time of their death. Also created a pretty cool particle system (housed in the Pickup Interface) that spawns when these souls are picked up by the Character. Added variables to the Attribute Component that links the number of gold and souls collected to display on the Character's HUD. Finally I added a dodge animation for the character to jump out the reach of the Enemy during combat. Dodging uses stamina (that regenerates at a specific rate), so there is a limit to the number of times the Character can dodge in any timeframe.